Refine expansion of function-typed placeholders
Per <https://github.com/swiftlang/sourcekit-lsp/pull/1789>, these
placeholders no longer expand to multi-line trailing closures. Instead
they become inline closures, with nested placeholders for the signature
and body.

This introduces a new formatter, `ClosureLiteralFormat`, so that other
uses of `BasicFormat` are not impacted. `ClosureLiteralFormat` does not
introduce newlines to a closure when there is only a single statement in
the body.